+++ demo/demo_form.php
@@ -48,6 +48,7 @@ EOT
     'req'	=>true,
     'check'	=>array("unique"),
     'order'	=>false,
+    'exclude_null_values' => true,
   ),
   'fav_hobby'   =>array(
     'type'      => 'select',
@@ -64,7 +65,7 @@ EOT
   'languages'	=>array(
     'name'	=>"What languages do you speak?",
     'type'	=>"checkbox",
-    'values'	=>array("php"=>"PHP", "perl"=>"Perl", "c"=>"C", "cpp"=>"C++", "java"=>"Java", "js"=>"Javascript"),
+    'values'	=>array("php"=>"PHP", "perl"=>"Perl", "c"=>"C", "cpp"=>"C++", "java"=>"Java", "js"=>array("name"=>"Javascript", "desc"=>"Script language, supported by web browsers")),
   ),
   'comment'	=>array(
     'name'	=>"Comment",